Create new doc using CSV in specified parent
The upload CSV method is working, but trying to append a note to these
documents causes an infinite loop of append>update headers>repeat.
Error msg: Invalid query parameter value for grid_id.
It appears that this may be a sheet issue. The uploaded files have a
sheet id that isn’t 0.