Stability and Validation
Fixed the validation of the form when creating a new document. Added an
XML sanitizer for catching unsupported characters when submitting via
the Spreadsheet API. This isn’t a problem with the Drive API. (TODO:
Can we move over entirely?) Added error handling for various cases like
when citable is triggered on chrome-*:// pages. (TODO: It’s a bit ugly,
we might want to use a regex in the future for page url matching. )
Also, finished the new doc flow, so that if the window stays open the
doclist gets updated and the form is reset. Clears the summary on
submit success to enable rapid note taking.