Cleanup and fixed button submit key behavior
There is a problem with using and input type=“submit” as the save
button- hitting return on the select triggers the submit. This is not
acceptable because you might be hitting return to try and change the
file and instead it submits. There should be a way to prevent this
which would be better for accessibility since we would keep the
type=“submit” for readers. Instead I’m using and anchor tag which is ok
and solves the problem.