Moved folder name to the shared props
This is for easy reference.

BUG: When switching folders, the doc menu defaults to Create New
Document but the new doc field doesn’t display. It’s easy to get out of
this, not fatal. But it’s not ideal. Should default to the first doc.
Need to check if the defaultId is in the list of docs and reset if not.
popup.js:423