Working Sheets UI
Update the Sheets UI to use an SVG data uri icon.
Fixed the message handling so that it triggers the auth loop when the
sheets UI is the first button pressed after install.
Correctly using google.storage.local for passing row and doc title info
to print and export. TODO: move this functionality to these pages so
that the data is cleaned up on tab close and each instance can contain
different data. Consider using url params to pass in the target doc id.
Updated the check controls to properly initialize and store their state
in localStorage. Should use the new storage APIs but I suppose it’s
fine. Not worth updating.
And fixed a display bug where Tags weren’t displaying properly in the
print view.