Updated the auth and online
Updated the auth design. and visual tweaks.
Fixed the online/offline keyboard behavior and note loading glitches.
Works even if the popup remains open, even though the auth flow will
close the popup.
I think the init display behavior is fixed, but there might still be
glitches. We’ll have to see.
Also added a fun feature for video tagging, it grabs the current time
from YouTube and puts it in the tags.
TODO: add support for vimeo, etc. & can we update the tags on
ctrl-enter for faster note taking? (it’ll get even better when we add
queueing).